Commit before fully breaking up the CLI
