+++ cli/__init__.py
@@ -1,0 +1,17 @@
+import logging
+import elasticsearch
+
+from ._version import __version__
+
+logger = logging.getLogger(__name__)
+
+from .utils import *
+from .cli import *
+from .alias import *
+from .allocate import *
+from .bloom import *
+from .close import *
+from .delete import *
+from .optimize import *
+from .replicas import *
+from .snapshot import *
